refactor(ka): remove dead impersonation code, update stale comments (#1288)
F-08: Delete CheckImpersonatePermission, NewImpersonatingConfig,
K8sCallAuditorImpl (zero production callers). Extract
InteractiveReadiness to its own file. Update event messages to
reflect MCP mount, not SSAR.
F-09: Update interfaces.go, server.go comments to remove
impersonation references.

FedRAMP: CM-6
